+++ web/src/components/GroupedChapterList.jsx
@@ -3,7 +3,7 @@ import React, { useEffect, useState } from 'react';
 import { Paper, Container, Typography, IconButton } from '@mui/material';
 
 import OpenInNewIcon from '@mui/icons-material/OpenInNew';
-import { groupBy } from '../utils/utilities';
+import { groupBy } from '../../common/utilities';
 import { formatChapterTitle, formatChapterUrl } from '../utils/formatting';
 import { MangaCover } from './MangaCover';
