+++ web/tsconfig.server.json
@@ -16,5 +16,11 @@
     "node_modules",
     "src",
     "__tests__/**"
-  ]
+  ],
+  "paths": {
+    "@/serverUtils*": ["utils*"],
+    "@/common*": ["common*"],
+    "@/db*": ["db*"],
+    "@/types*": ["types*"]
+  }
 }
